Puppet.lns cannot parse this kind of key [replication_port://8079]
Replication port may be configured in server.conf as documented on
https://docs.splunk.com/Documentation/Splunk/8.2.6/Indexer/Configurepeerswithserverconf
but the presence of this stanza prevents augeas from being able to parse
the file.

The "Splunk.lns" is available in Puppet 5.5, but comes with a
few caveats regarding its usage

https://github.com/hercules-team/augeas/commit/b64ee52dbcc5c5fdfd31f3a6bf13f33d98f2e809

Lens to parse Splunk configuration. Based on IniFile lens, with few
differences:

* Accepts empty value in entry
* Entries without parent section are children of .anon node